refactor: move error to dedicated module
